+++ EasyPost/Parameters/Tracker/All.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using EasyPost.Utilities.Internal.Attributes;
@@ -51,9 +52,18 @@ namespace EasyPost.Parameters.Tracker
 
         /// <summary>
         ///     Only return trackers with the given tracking code.
+        ///     Deprecated: Use <see cref="TrackingCodes"/> instead.
         /// </summary>
+        // TODO: Remove in next major version.
+        [Obsolete("This property will be removed in a future version and replaced with TrackingCodes.")]
         [TopLevelRequestParameter(Necessity.Optional, "tracking_code")]
         public string? TrackingCode { get; set; }
+
+        /// <summary>
+        ///     Only return trackers with the given tracking codes.
+        /// </summary>
+        [TopLevelRequestParameter(Necessity.Optional, "tracking_codes")]
+        public List<string>? TrackingCodes { get; set; }
 
         #endregion